Fix: charge for #looting pet in a shop
Taking an item from your pet while inside a shop would charge you for
the item, even if it wasn't owned by the shop.  One particularly
illustrative case is a scenario where the hero is standing in the shop
doorway, her pet is in the hallway outside, and the item in question has
never even been inside the shop -- it would still be added to the bill,
and since the hero was already on the shop threshold, a typo could
easily anger the shopkeeper and summon the kops.

Account for the 'free spot(s)' inside a shop (the doorway and the spot
immediately next to it) when considering whether to charge for an item
taken from a pet.  Mark items given to a pet no-charge if they are
already the hero's property, so that giving an item to a pet and
immediately taking it back while standing in a shop, even on a 'costly
square', doesn't cost money.

Also include the '(no charge)' slug for objects in a pet's inventory:
otherwise, even uncharged items would be marked as 'for sale'.  This
description was previously only applied to items on the floor.