Do charge for items given to pets in a shop
I don't mind the idea that giving an item to a pet marks it as
'no-charge', so that even if the pet drops it you won't lose it to the
shop's inventory.  I think the alternative (the current state of
affairs) is a very very strong disincentive from ever giving an item to
your pet while inside a shop, and the fact that the item becomes the
shop's property as soon as it enters your pet's inventory compounds that
feeling.

But in the hopes of at least saving the "don't charge for items a pet
brings to you in the hallway while you stand in the shop door" bit of
this PR, I'll begrudgingly acquiesce on this point. ;)

This is a separate commit purely so that it's clearer how I updated the
PR -- I can squash everything together if needed, etc.